refactor: further split editing and rendering into focused modules
Continue splitting rich_text_editor into even more focused modules
following single responsibility principle.

## New Editing Modules (from editing.rs)

### cursor.rs (81 lines)
- Cursor movement in all 8 directions
- move_left/right/up/down, move_home/end, move_document_start/end

### selection.rs (107 lines)
- Selection handling: start, clear, has, get, delete
- Multi-block selection support

### text_edit.rs (173 lines)
- Text insertion/deletion
- Block operations: split, merge (with previous/next), delete
- Private helper functions

### format.rs (58 lines)
- Formatting toggles: bold, italic, underline, strikethrough, code
- Block type management: set_block_type, current_block_type
- Current format accessor

### undo.rs (105 lines)
- MAX_UNDO_HISTORY constant
- Undo/redo stack management
- Push operations with history limit

## New Rendering Modules (from rendering.rs)

### toolbar.rs (62 lines)
- Toolbar rendering with active states
- Toolbar items: formatting, block types, link/image

### editor_view.rs (120 lines)
- Editor view rendering with cursor
- Block type indicators and prefixes
- Selection highlighting

### preview.rs (169 lines)
- Preview mode rendering with formatted text
- Block-specific rendering (headings, quotes, lists, code)
- render_heading() helper for heading text

### dialog.rs (180 lines)
- Insert link/image dialog UI
- Border drawing, title, input fields
- Dialog positioning and layout

## Updated Files

### editing.rs (188 lines, reduced from 694 lines)
- Kept: toolbar_action, process_markdown_shortcuts, apply_shortcut
- Kept: handle_key, handle_dialog_key (orchestration)

### mod.rs (1,301 lines, reduced from 1,293 lines)
- Added module declarations for all 9 new modules
- Each module has a single, clear responsibility

## Benefits

- **Single Responsibility**: Each module does ONE thing well
- **Easy Navigation**: Find functionality instantly by module name
- **Parallel Development**: Multiple developers can work on different modules
- **Focused Testing**: Test files can target specific modules
- **Reduced Cognitive Load**: Smaller files are easier to understand

## Total Modules: 17 files

- Core: core.rs, editing.rs
- Editing: cursor.rs, selection.rs, text_edit.rs, format.rs, undo.rs
- Rendering: toolbar.rs, editor_view.rs, preview.rs, dialog.rs
- Support: markdown.rs, block.rs, link.rs, text_format.rs, types.rs
- Main: mod.rs (tests + View impl + enums)

Fixes #247
